backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
psycopg2
python-jose[cryptography]
passlib
bcrypt==4.0.1
httpx
python-multipart
pydantic[email]
python-dotenv
email-validator
